% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrationCurve.R
\name{dsCalibrationCurve}
\alias{dsCalibrationCurve}
\title{Calculate calibration curve on DataSHIELD servers}
\usage{
dsCalibrationCurve(
  connections,
  truth_name,
  pred_name,
  nbins = 10L,
  nprivacy = 5L
)
}
\arguments{
\item{connections}{[DSI::connection] Connection to an OPAL server.}

\item{truth_name}{[character(1L)] R object containing the models name as character.}

\item{pred_name}{[character(1L)] Name of the object predictions should be assigned to.}

\item{nbins}{[integer(1L)] Number of bins between 0 and 1 (default is 10).}

\item{nprivacy}{[integer(1L)] Number of values per bin to ensure privacy (default is 5).}
}
\value{
Brier score for multiple server
}
\description{
This function enables to calculate a calibration curve on DataSHIELD servers.
}
\author{
Daniel S.
}
