% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrationCurve.R
\name{plotCalibrationCurve}
\alias{plotCalibrationCurve}
\title{Plot calibration curve}
\usage{
plotCalibrationCurve(cc, individuals = TRUE)
}
\arguments{
\item{cc}{(\code{list()}) Object returned by \code{dsCalibrationCurve()}}

\item{individuals}{(\code{logical(1L)}) Logical value indicating whether the individual calibration
curves should be plotted or not (default is \code{TRUE}).}
}
\value{
ggplot of calibration curve(s)
}
\description{
This function plots the calibration curve returned by \code{dsCalibrationCurve()}.
}
\author{
Daniel S.
}
